requirements_gui.txt
PyQt6>=6.4.0
pywinpty